# add steiner support and add test cases from earcut.js