--- a/test/fixtures/issue45.cpp
+++ b/test/fixtures/issue45.cpp
@@ -1,0 +1,13 @@
+#include "geometries.hpp"
+
+namespace mapbox {
+namespace fixtures {
+
+const IntegerPolygon issue45 = {
+    {{10,10}, {25,10}, {25,40}, {10,40}},
+    {{15,30}, {20,35}, {10,40}},
+    {{15,15}, {15,20}, {20,15}}
+};
+
+}
+}