--- a/test/fixtures/steiner.cpp
+++ b/test/fixtures/steiner.cpp
@@ -1,0 +1,15 @@
+#include "geometries.hpp"
+
+namespace mapbox {
+namespace fixtures {
+
+const IntegerPolygon steiner = {
+    {{0,0},{100,0},{100,100},{0,100}},
+    {{50,50}},
+    {{30,40}},
+    {{70,60}},
+    {{20,70}}
+};
+
+}
+}