--- a/test/test.cpp
+++ b/test/test.cpp
@@ -112,32 +112,32 @@ int main() {
         t.end();
     }
 
-    areaTest<int>("building", mapbox::fixtures::building, 12);
+    areaTest<int>("building", mapbox::fixtures::building, 13);
     areaTest<double>("dude", mapbox::fixtures::dude, 106);
     areaTest<int>("water", mapbox::fixtures::water, 2482, 0.0008, 0.00002);
-    areaTest<int>("water2", mapbox::fixtures::water2, 1211);
+    areaTest<int>("water2", mapbox::fixtures::water2, 1212);
     areaTest<int>("water3", mapbox::fixtures::water3, 197);
     areaTest<int>("water3b", mapbox::fixtures::water3b, 25);
     areaTest<int>("water4", mapbox::fixtures::water4, 705);
     areaTest<int>("water_huge", mapbox::fixtures::water_huge, 5159, 0.008, 0.0002);
-    areaTest<int>("water_huge2", mapbox::fixtures::water_huge2, 4458, 0.0019, 0.00015);
+    areaTest<int>("water_huge2", mapbox::fixtures::water_huge2, 4456, 0.0019, 0.00015);
     areaTest<int>("degenerate", mapbox::fixtures::degenerate, 0);
-    areaTest<int>("bad_hole", mapbox::fixtures::bad_hole, 34, 0.042, 0.0022);
+    areaTest<int>("bad_hole", mapbox::fixtures::bad_hole, 42, 0.042, 0.0022);
     // allow libtess2 failure on this by providing infinity.
     areaTest<int>("empty_square", mapbox::fixtures::empty_square, 0, 0, std::numeric_limits<double>::infinity());
-    areaTest<int>("issue16", mapbox::fixtures::issue16, 12);
-    areaTest<int>("issue17", mapbox::fixtures::issue17, 11);
+    areaTest<double>("issue16", mapbox::fixtures::issue16, 12, 1e-14, 0.0255);
+    areaTest<double>("issue17", mapbox::fixtures::issue17, 11);
     areaTest<int>("steiner", mapbox::fixtures::steiner, 9);
-    areaTest<int>("issue29", mapbox::fixtures::issue29, 40);
-    areaTest<int>("issue34", mapbox::fixtures::issue34, 139);
-    areaTest<int>("issue35", mapbox::fixtures::issue35, 844);
+    areaTest<double>("issue29", mapbox::fixtures::issue29, 40);
+    areaTest<int>("issue34", mapbox::fixtures::issue34, 138);
+    areaTest<int>("issue35", mapbox::fixtures::issue35, 841);
     areaTest<double>("self-touching", mapbox::fixtures::self_touching, 124, 3.4e-14, 0.002);
     areaTest<int>("outside-ring", mapbox::fixtures::outside_ring, 64);
     areaTest<int>("simplified-us-border", mapbox::fixtures::simplified_us_border, 120, 1e-14, 0.001);
     areaTest<int>("touching-holes", mapbox::fixtures::touching_holes, 57);
     areaTest<int>("hole-touching-outer", mapbox::fixtures::hole_touching_outer, 77);
     areaTest<int>("hilbert", mapbox::fixtures::hilbert, 1023);
-    areaTest<int>("issue45", mapbox::fixtures::issue45, 10);
+    areaTest<int>("issue45", mapbox::fixtures::issue45, 10, 1e-14, 0.094);
     areaTest<short>("park", mapbox::fixtures::park);
 
     return 0;