Port of steiner points (see https://github.com/mapbox/earcut.hpp/pull/23#issuecomment-182939174).
Also added missing tests from earcut.js