@mourner 2016-02-10
Here's the reason why we haven't yet ported the tests: https://github.com/mapbox/earcut.hpp/issues/6#issuecomment-147784634

If you find out why the tests fail and fix the PR so that the build passes, that would awesome!