@mrgreywater 2016-02-10
yes, this PR is based on the port-tests branch, because there were still new tests in earcut.js that were missing in this branch.