@mourner 2016-02-10
@mrgreywater but the build on Travis still fails for this PR, right? How could this be fixed?