@mrgreywater 2016-02-10
@mourner If I'm reading the output right the build completes and runs, but the tests are failing, so travis reports that the checks failed. This is fine, since the whole purpose of this branch is that the tests may fail.