@mourner 2016-02-11
@mrgreywater if the tests are failing, Travis looses its sense since we won't be able to track down regressions. We need to figure out how to make them pass before we can merge new tests.