@mrgreywater 2016-02-11
this is adds https://github.com/mapbox/earcut/commit/deb0bfee2245012c98f62f1ee3bb6a206b89db1e https://github.com/mapbox/earcut/commit/65c7da544996fb479a2caa3d4188734cb22ce5d8 https://github.com/mapbox/earcut/commit/68da53f814c3ab2ad85aebf02f46fd48f0605421 and some others. Test cases now match earcut.js again.