@mrgreywater 2016-02-11
Before:
```
+----------------+--------------------+--------------------+
| Polygon        | earcut             | libtess2           |
+----------------+--------------------+--------------------+
| bad_hole       |      307.127 ops/s |       21.730 ops/s |
| building       |    1.764.854 ops/s |       51.918 ops/s |
| degenerate     |    4.069.767 ops/s |       90.989 ops/s |
| dude           |      103.678 ops/s |       11.735 ops/s |
| empty_square   |    2.673.845 ops/s |       78.770 ops/s |
| water_huge     |          218 ops/s |           60 ops/s |
| water_huge2    |           87 ops/s |           76 ops/s |
| water          |        1.288 ops/s |          157 ops/s |
| water2         |        1.476 ops/s |          715 ops/s |
| water3         |       43.150 ops/s |        5.477 ops/s |
| water3b        |      473.830 ops/s |       32.914 ops/s |
| water4         |        5.412 ops/s |        1.305 ops/s |
+----------------+--------------------+--------------------+
```
After:
```
+----------------+--------------------+--------------------+
| Polygon        | earcut             | libtess2           |
+----------------+--------------------+--------------------+
| bad_hole       |      175.355 ops/s |       21.754 ops/s |
| building       |    1.899.877 ops/s |       50.988 ops/s |
| degenerate     |    3.422.676 ops/s |       89.196 ops/s |
| dude           |      107.021 ops/s |       11.673 ops/s |
| empty_square   |    2.373.604 ops/s |       76.254 ops/s |
| water_huge     |           87 ops/s |           59 ops/s |
| water_huge2    |           42 ops/s |           78 ops/s |
| water          |        1.192 ops/s |          157 ops/s |
| water2         |        1.333 ops/s |          712 ops/s |
| water3         |       38.208 ops/s |        5.436 ops/s |
| water3b        |      409.587 ops/s |       32.757 ops/s |
| water4         |        4.874 ops/s |        1.297 ops/s |
+----------------+--------------------+--------------------+
```
as @mourner said [here
](https://github.com/mapbox/earcut.hpp/issues/6#issuecomment-168670243) https://github.com/mapbox/earcut/commit/deb0bfee2245012c98f62f1ee3bb6a206b89db1e causes some performance regression, but unless we want to diverge both in the earcut.hpp and in the unit-tests, there is not much we can do about it. I'm curious about the performance difference that this commit had to earcut.js.