@mourner 2016-02-11
This is great, thanks for bring earcut.hpp up to date and fixing tests! The regression from my benchmarks looked a bit worse as far as I remember but that may be just some platform deviation, and after all I think it only regresses on really bad data with lots of self-intersections. When the polygons are clean, the performance should improve (like the building case above). So I'll merge this to master.