@mourner 2016-02-12
Rebased and merged to master :+1: