@mrgreywater 2016-02-12
Mmh, that rebase was kind of unnecessary. I would've just merged it onto port-tests and then merged that branch with master. Thanks for adding the changes though.