@mourner 2016-02-12
@mrgreywater just reducing the number of merge commits in git history, and rebase was clean.